slice: scale dragon and use existing zone asset
- Default zone -> assets/models/ruins.gltf to avoid missing path
- Spawn dragon with small scale and offset so camera isn’t inside
- Stop double-initializing logger (remove env_logger init)
