+++ Cargo.toml
@@ -38,6 +38,7 @@ platform_winit = { version = "0.1.0", path = "crates/platform_winit", default-fe
 ecs_core = { version = "0.1.0", path = "crates/ecs_core" }
 client_core = { version = "0.1.0", path = "crates/client_core" }
 sim_core = { version = "0.1.0", path = "crates/sim_core" }
+roa_slice_bevy = { version = "0.1.0", path = "apps/roa_slice_bevy" }
 
 [patch.crates-io]
 # Use a patched local copy of draco_decoder to avoid passing
